Update SVG build script to set custom dimensions
This removes the fixed width and height override from the build script and instead detects the viewBox values and sets the width and height from there. The goal of this is allow for different aspect ratios in icons. My gut says everything should be 16px tall to address vertical align issues, but then have some variable width.

To start, this will be for select icons (possibly the new Bootstrap logo and our upcoming badge icons), and then potentially redoing artboards in Figma for super narrow icons so they only take up the required space. That'll come later I think though.